requirements.txt
Django==3.1
Pillow==8.1.1
djangorestframework==3.11
python-decouple==3.4
drf-yasg
psycopg2-binary==2.8.6
BeautifulSoup4
requests
lxml
